requirements.txt
intersystems-irispython
pydantic
langchain-core
langchain_ollama
streamlit
streamlit-chat
